backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.0",
    "fastapi>=0.129.0",
    "firebase-admin>=7.1.0",
    "google-generativeai>=0.8.6",
    "pdfplumber>=0.11.9",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "supabase>=2.28.0",
    "uvicorn[standard]>=0.40.0",
]
